CodeGen from PR 24880 in Azure/azure-rest-api-specs Merge 518c270ba5771fd355e08d98969194a524442e3f into 1af2861030243b06ee35172c95899f4809eedfc7
