+++ sdk/eventgrid/azure-resourcemanager-eventgrid/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release History
 
-## 1.2.0-beta.5 (Unreleased)
+## 1.0.0-beta.1 (2023-07-18)
+
+- Azure Resource Manager EventGrid client library for Java. This package contains Microsoft Azure SDK for EventGrid Management SDK. Azure EventGrid Management Client. Package tag package-2023-06-preview. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 
 ### Features Added
