+++ sdk/eventgrid/azure-resourcemanager-eventgrid/src/main/java/com/azure/resourcemanager/eventgrid/models/Channel.java
@@ -112,11 +112,13 @@ public interface Channel {
     interface Definition
         extends DefinitionStages.Blank, DefinitionStages.WithParentResource, DefinitionStages.WithCreate {
     }
+
     /** The Channel definition stages. */
     interface DefinitionStages {
         /** The first stage of the Channel definition. */
         interface Blank extends WithParentResource {
         }
+
         /** The stage of the Channel definition allowing to specify parent resource. */
         interface WithParentResource {
             /**
@@ -128,6 +130,7 @@ public interface Channel {
              */
             WithCreate withExistingPartnerNamespace(String resourceGroupName, String partnerNamespaceName);
         }
+
         /**
          * The stage of the Channel definition which contains all the minimum required properties for the resource to be
          * created, but also allows for any other optional properties to be specified.
@@ -155,6 +158,7 @@ public interface Channel {
              */
             Channel create(Context context);
         }
+
         /** The stage of the Channel definition allowing to specify channelType. */
         interface WithChannelType {
             /**
@@ -166,6 +170,7 @@ public interface Channel {
              */
             WithCreate withChannelType(ChannelType channelType);
         }
+
         /** The stage of the Channel definition allowing to specify partnerTopicInfo. */
         interface WithPartnerTopicInfo {
             /**
@@ -178,6 +183,7 @@ public interface Channel {
              */
             WithCreate withPartnerTopicInfo(PartnerTopicInfo partnerTopicInfo);
         }
+
         /** The stage of the Channel definition allowing to specify partnerDestinationInfo. */
         interface WithPartnerDestinationInfo {
             /**
@@ -191,6 +197,7 @@ public interface Channel {
              */
             WithCreate withPartnerDestinationInfo(PartnerDestinationInfo partnerDestinationInfo);
         }
+
         /** The stage of the Channel definition allowing to specify messageForActivation. */
         interface WithMessageForActivation {
             /**
@@ -203,6 +210,7 @@ public interface Channel {
              */
             WithCreate withMessageForActivation(String messageForActivation);
         }
+
         /** The stage of the Channel definition allowing to specify provisioningState. */
         interface WithProvisioningState {
             /**
@@ -213,6 +221,7 @@ public interface Channel {
              */
             WithCreate withProvisioningState(ChannelProvisioningState provisioningState);
         }
+
         /** The stage of the Channel definition allowing to specify readinessState. */
         interface WithReadinessState {
             /**
@@ -223,6 +232,7 @@ public interface Channel {
              */
             WithCreate withReadinessState(ReadinessState readinessState);
         }
+
         /** The stage of the Channel definition allowing to specify expirationTimeIfNotActivatedUtc. */
         interface WithExpirationTimeIfNotActivatedUtc {
             /**
@@ -238,6 +248,7 @@ public interface Channel {
             WithCreate withExpirationTimeIfNotActivatedUtc(OffsetDateTime expirationTimeIfNotActivatedUtc);
         }
     }
+
     /**
      * Begins update for the Channel resource.
      *
@@ -269,6 +280,7 @@ public interface Channel {
          */
         Channel apply(Context context);
     }
+
     /** The Channel update stages. */
     interface UpdateStages {
         /** The stage of the Channel update allowing to specify channelType. */
@@ -282,6 +294,7 @@ public interface Channel {
              */
             Update withChannelType(ChannelType channelType);
         }
+
         /** The stage of the Channel update allowing to specify partnerTopicInfo. */
         interface WithPartnerTopicInfo {
             /**
@@ -294,6 +307,7 @@ public interface Channel {
              */
             Update withPartnerTopicInfo(PartnerTopicInfo partnerTopicInfo);
         }
+
         /** The stage of the Channel update allowing to specify partnerDestinationInfo. */
         interface WithPartnerDestinationInfo {
             /**
@@ -307,6 +321,7 @@ public interface Channel {
              */
             Update withPartnerDestinationInfo(PartnerDestinationInfo partnerDestinationInfo);
         }
+
         /** The stage of the Channel update allowing to specify messageForActivation. */
         interface WithMessageForActivation {
             /**
@@ -319,6 +334,7 @@ public interface Channel {
              */
             Update withMessageForActivation(String messageForActivation);
         }
+
         /** The stage of the Channel update allowing to specify provisioningState. */
         interface WithProvisioningState {
             /**
@@ -329,6 +345,7 @@ public interface Channel {
              */
             Update withProvisioningState(ChannelProvisioningState provisioningState);
         }
+
         /** The stage of the Channel update allowing to specify readinessState. */
         interface WithReadinessState {
             /**
@@ -339,6 +356,7 @@ public interface Channel {
              */
             Update withReadinessState(ReadinessState readinessState);
         }
+
         /** The stage of the Channel update allowing to specify expirationTimeIfNotActivatedUtc. */
         interface WithExpirationTimeIfNotActivatedUtc {
             /**
@@ -354,6 +372,7 @@ public interface Channel {
             Update withExpirationTimeIfNotActivatedUtc(OffsetDateTime expirationTimeIfNotActivatedUtc);
         }
     }
+
     /**
      * Refreshes the resource to sync with Azure.
      *
