+++ sdk/eventgrid/azure-resourcemanager-eventgrid/src/main/java/com/azure/resourcemanager/eventgrid/models/Domain.java
@@ -228,11 +228,13 @@ public interface Domain {
             DefinitionStages.WithResourceGroup,
             DefinitionStages.WithCreate {
     }
+
     /** The Domain definition stages. */
     interface DefinitionStages {
         /** The first stage of the Domain definition. */
         interface Blank extends WithLocation {
         }
+
         /** The stage of the Domain definition allowing to specify location. */
         interface WithLocation {
             /**
@@ -251,6 +253,7 @@ public interface Domain {
              */
             WithResourceGroup withRegion(String location);
         }
+
         /** The stage of the Domain definition allowing to specify parent resource. */
         interface WithResourceGroup {
             /**
@@ -261,6 +264,7 @@ public interface Domain {
              */
             WithCreate withExistingResourceGroup(String resourceGroupName);
         }
+
         /**
          * The stage of the Domain definition which contains all the minimum required properties for the resource to be
          * created, but also allows for any other optional properties to be specified.
@@ -294,6 +298,7 @@ public interface Domain {
              */
             Domain create(Context context);
         }
+
         /** The stage of the Domain definition allowing to specify tags. */
         interface WithTags {
             /**
@@ -304,6 +309,7 @@ public interface Domain {
              */
             WithCreate withTags(Map<String, String> tags);
         }
+
         /** The stage of the Domain definition allowing to specify sku. */
         interface WithSku {
             /**
@@ -314,6 +320,7 @@ public interface Domain {
              */
             WithCreate withSku(ResourceSku sku);
         }
+
         /** The stage of the Domain definition allowing to specify identity. */
         interface WithIdentity {
             /**
@@ -324,6 +331,7 @@ public interface Domain {
              */
             WithCreate withIdentity(IdentityInfo identity);
         }
+
         /** The stage of the Domain definition allowing to specify minimumTlsVersionAllowed. */
         interface WithMinimumTlsVersionAllowed {
             /**
@@ -335,6 +343,7 @@ public interface Domain {
              */
             WithCreate withMinimumTlsVersionAllowed(TlsVersion minimumTlsVersionAllowed);
         }
+
         /** The stage of the Domain definition allowing to specify inputSchema. */
         interface WithInputSchema {
             /**
@@ -347,6 +356,7 @@ public interface Domain {
              */
             WithCreate withInputSchema(InputSchema inputSchema);
         }
+
         /** The stage of the Domain definition allowing to specify eventTypeInfo. */
         interface WithEventTypeInfo {
             /**
@@ -359,6 +369,7 @@ public interface Domain {
              */
             WithCreate withEventTypeInfo(EventTypeInfo eventTypeInfo);
         }
+
         /** The stage of the Domain definition allowing to specify inputSchemaMapping. */
         interface WithInputSchemaMapping {
             /**
@@ -371,6 +382,7 @@ public interface Domain {
              */
             WithCreate withInputSchemaMapping(InputSchemaMapping inputSchemaMapping);
         }
+
         /** The stage of the Domain definition allowing to specify publicNetworkAccess. */
         interface WithPublicNetworkAccess {
             /**
@@ -386,6 +398,7 @@ public interface Domain {
              */
             WithCreate withPublicNetworkAccess(PublicNetworkAccess publicNetworkAccess);
         }
+
         /** The stage of the Domain definition allowing to specify inboundIpRules. */
         interface WithInboundIpRules {
             /**
@@ -398,6 +411,7 @@ public interface Domain {
              */
             WithCreate withInboundIpRules(List<InboundIpRule> inboundIpRules);
         }
+
         /** The stage of the Domain definition allowing to specify disableLocalAuth. */
         interface WithDisableLocalAuth {
             /**
@@ -412,6 +426,7 @@ public interface Domain {
              */
             WithCreate withDisableLocalAuth(Boolean disableLocalAuth);
         }
+
         /** The stage of the Domain definition allowing to specify autoCreateTopicWithFirstSubscription. */
         interface WithAutoCreateTopicWithFirstSubscription {
             /**
@@ -440,6 +455,7 @@ public interface Domain {
              */
             WithCreate withAutoCreateTopicWithFirstSubscription(Boolean autoCreateTopicWithFirstSubscription);
         }
+
         /** The stage of the Domain definition allowing to specify autoDeleteTopicWithLastSubscription. */
         interface WithAutoDeleteTopicWithLastSubscription {
             /**
@@ -468,6 +484,7 @@ public interface Domain {
              */
             WithCreate withAutoDeleteTopicWithLastSubscription(Boolean autoDeleteTopicWithLastSubscription);
         }
+
         /** The stage of the Domain definition allowing to specify dataResidencyBoundary. */
         interface WithDataResidencyBoundary {
             /**
@@ -479,6 +496,7 @@ public interface Domain {
             WithCreate withDataResidencyBoundary(DataResidencyBoundary dataResidencyBoundary);
         }
     }
+
     /**
      * Begins update for the Domain resource.
      *
@@ -514,6 +532,7 @@ public interface Domain {
          */
         Domain apply(Context context);
     }
+
     /** The Domain update stages. */
     interface UpdateStages {
         /** The stage of the Domain update allowing to specify tags. */
@@ -526,6 +545,7 @@ public interface Domain {
              */
             Update withTags(Map<String, String> tags);
         }
+
         /** The stage of the Domain update allowing to specify identity. */
         interface WithIdentity {
             /**
@@ -536,6 +556,7 @@ public interface Domain {
              */
             Update withIdentity(IdentityInfo identity);
         }
+
         /** The stage of the Domain update allowing to specify sku. */
         interface WithSku {
             /**
@@ -546,6 +567,7 @@ public interface Domain {
              */
             Update withSku(ResourceSku sku);
         }
+
         /** The stage of the Domain update allowing to specify publicNetworkAccess. */
         interface WithPublicNetworkAccess {
             /**
@@ -562,6 +584,7 @@ public interface Domain {
              */
             Update withPublicNetworkAccess(PublicNetworkAccess publicNetworkAccess);
         }
+
         /** The stage of the Domain update allowing to specify inboundIpRules. */
         interface WithInboundIpRules {
             /**
@@ -574,6 +597,7 @@ public interface Domain {
              */
             Update withInboundIpRules(List<InboundIpRule> inboundIpRules);
         }
+
         /** The stage of the Domain update allowing to specify minimumTlsVersionAllowed. */
         interface WithMinimumTlsVersionAllowed {
             /**
@@ -585,6 +609,7 @@ public interface Domain {
              */
             Update withMinimumTlsVersionAllowed(TlsVersion minimumTlsVersionAllowed);
         }
+
         /** The stage of the Domain update allowing to specify disableLocalAuth. */
         interface WithDisableLocalAuth {
             /**
@@ -599,6 +624,7 @@ public interface Domain {
              */
             Update withDisableLocalAuth(Boolean disableLocalAuth);
         }
+
         /** The stage of the Domain update allowing to specify autoCreateTopicWithFirstSubscription. */
         interface WithAutoCreateTopicWithFirstSubscription {
             /**
@@ -627,6 +653,7 @@ public interface Domain {
              */
             Update withAutoCreateTopicWithFirstSubscription(Boolean autoCreateTopicWithFirstSubscription);
         }
+
         /** The stage of the Domain update allowing to specify autoDeleteTopicWithLastSubscription. */
         interface WithAutoDeleteTopicWithLastSubscription {
             /**
@@ -655,6 +682,7 @@ public interface Domain {
              */
             Update withAutoDeleteTopicWithLastSubscription(Boolean autoDeleteTopicWithLastSubscription);
         }
+
         /** The stage of the Domain update allowing to specify dataResidencyBoundary. */
         interface WithDataResidencyBoundary {
             /**
@@ -665,6 +693,7 @@ public interface Domain {
              */
             Update withDataResidencyBoundary(DataResidencyBoundary dataResidencyBoundary);
         }
+
         /** The stage of the Domain update allowing to specify eventTypeInfo. */
         interface WithEventTypeInfo {
             /**
@@ -676,6 +705,7 @@ public interface Domain {
             Update withEventTypeInfo(EventTypeInfo eventTypeInfo);
         }
     }
+
     /**
      * Refreshes the resource to sync with Azure.
      *
