+++ sdk/eventgrid/azure-resourcemanager-eventgrid/src/main/java/com/azure/resourcemanager/eventgrid/models/Namespace.java
@@ -169,11 +169,13 @@ public interface Namespace {
             DefinitionStages.WithResourceGroup,
             DefinitionStages.WithCreate {
     }
+
     /** The Namespace definition stages. */
     interface DefinitionStages {
         /** The first stage of the Namespace definition. */
         interface Blank extends WithLocation {
         }
+
         /** The stage of the Namespace definition allowing to specify location. */
         interface WithLocation {
             /**
@@ -192,6 +194,7 @@ public interface Namespace {
              */
             WithResourceGroup withRegion(String location);
         }
+
         /** The stage of the Namespace definition allowing to specify parent resource. */
         interface WithResourceGroup {
             /**
@@ -202,6 +205,7 @@ public interface Namespace {
              */
             WithCreate withExistingResourceGroup(String resourceGroupName);
         }
+
         /**
          * The stage of the Namespace definition which contains all the minimum required properties for the resource to
          * be created, but also allows for any other optional properties to be specified.
@@ -232,6 +236,7 @@ public interface Namespace {
              */
             Namespace create(Context context);
         }
+
         /** The stage of the Namespace definition allowing to specify tags. */
         interface WithTags {
             /**
@@ -242,6 +247,7 @@ public interface Namespace {
              */
             WithCreate withTags(Map<String, String> tags);
         }
+
         /** The stage of the Namespace definition allowing to specify sku. */
         interface WithSku {
             /**
@@ -252,6 +258,7 @@ public interface Namespace {
              */
             WithCreate withSku(NamespaceSku sku);
         }
+
         /** The stage of the Namespace definition allowing to specify identity. */
         interface WithIdentity {
             /**
@@ -262,6 +269,7 @@ public interface Namespace {
              */
             WithCreate withIdentity(IdentityInfo identity);
         }
+
         /** The stage of the Namespace definition allowing to specify privateEndpointConnections. */
         interface WithPrivateEndpointConnections {
             /**
@@ -272,6 +280,7 @@ public interface Namespace {
              */
             WithCreate withPrivateEndpointConnections(List<PrivateEndpointConnectionInner> privateEndpointConnections);
         }
+
         /** The stage of the Namespace definition allowing to specify topicsConfiguration. */
         interface WithTopicsConfiguration {
             /**
@@ -282,6 +291,7 @@ public interface Namespace {
              */
             WithCreate withTopicsConfiguration(TopicsConfiguration topicsConfiguration);
         }
+
         /** The stage of the Namespace definition allowing to specify topicSpacesConfiguration. */
         interface WithTopicSpacesConfiguration {
             /**
@@ -293,6 +303,7 @@ public interface Namespace {
              */
             WithCreate withTopicSpacesConfiguration(TopicSpacesConfiguration topicSpacesConfiguration);
         }
+
         /** The stage of the Namespace definition allowing to specify isZoneRedundant. */
         interface WithIsZoneRedundant {
             /**
@@ -307,6 +318,7 @@ public interface Namespace {
              */
             WithCreate withIsZoneRedundant(Boolean isZoneRedundant);
         }
+
         /** The stage of the Namespace definition allowing to specify publicNetworkAccess. */
         interface WithPublicNetworkAccess {
             /**
@@ -323,6 +335,7 @@ public interface Namespace {
              */
             WithCreate withPublicNetworkAccess(PublicNetworkAccess publicNetworkAccess);
         }
+
         /** The stage of the Namespace definition allowing to specify inboundIpRules. */
         interface WithInboundIpRules {
             /**
@@ -335,6 +348,7 @@ public interface Namespace {
              */
             WithCreate withInboundIpRules(List<InboundIpRule> inboundIpRules);
         }
+
         /** The stage of the Namespace definition allowing to specify minimumTlsVersionAllowed. */
         interface WithMinimumTlsVersionAllowed {
             /**
@@ -348,6 +362,7 @@ public interface Namespace {
             WithCreate withMinimumTlsVersionAllowed(TlsVersion minimumTlsVersionAllowed);
         }
     }
+
     /**
      * Begins update for the Namespace resource.
      *
@@ -378,6 +393,7 @@ public interface Namespace {
          */
         Namespace apply(Context context);
     }
+
     /** The Namespace update stages. */
     interface UpdateStages {
         /** The stage of the Namespace update allowing to specify tags. */
@@ -390,6 +406,7 @@ public interface Namespace {
              */
             Update withTags(Map<String, String> tags);
         }
+
         /** The stage of the Namespace update allowing to specify identity. */
         interface WithIdentity {
             /**
@@ -400,6 +417,7 @@ public interface Namespace {
              */
             Update withIdentity(IdentityInfo identity);
         }
+
         /** The stage of the Namespace update allowing to specify sku. */
         interface WithSku {
             /**
@@ -410,6 +428,7 @@ public interface Namespace {
              */
             Update withSku(NamespaceSku sku);
         }
+
         /** The stage of the Namespace update allowing to specify topicSpacesConfiguration. */
         interface WithTopicSpacesConfiguration {
             /**
@@ -421,6 +440,7 @@ public interface Namespace {
              */
             Update withTopicSpacesConfiguration(UpdateTopicSpacesConfigurationInfo topicSpacesConfiguration);
         }
+
         /** The stage of the Namespace update allowing to specify publicNetworkAccess. */
         interface WithPublicNetworkAccess {
             /**
@@ -437,6 +457,7 @@ public interface Namespace {
              */
             Update withPublicNetworkAccess(PublicNetworkAccess publicNetworkAccess);
         }
+
         /** The stage of the Namespace update allowing to specify inboundIpRules. */
         interface WithInboundIpRules {
             /**
@@ -450,6 +471,7 @@ public interface Namespace {
             Update withInboundIpRules(List<InboundIpRule> inboundIpRules);
         }
     }
+
     /**
      * Refreshes the resource to sync with Azure.
      *
