+++ sdk/eventgrid/azure-resourcemanager-eventgrid/src/main/java/com/azure/resourcemanager/eventgrid/models/PartnerDestination.java
@@ -143,11 +143,13 @@ public interface PartnerDestination {
             DefinitionStages.WithResourceGroup,
             DefinitionStages.WithCreate {
     }
+
     /** The PartnerDestination definition stages. */
     interface DefinitionStages {
         /** The first stage of the PartnerDestination definition. */
         interface Blank extends WithLocation {
         }
+
         /** The stage of the PartnerDestination definition allowing to specify location. */
         interface WithLocation {
             /**
@@ -166,6 +168,7 @@ public interface PartnerDestination {
              */
             WithResourceGroup withRegion(String location);
         }
+
         /** The stage of the PartnerDestination definition allowing to specify parent resource. */
         interface WithResourceGroup {
             /**
@@ -176,6 +179,7 @@ public interface PartnerDestination {
              */
             WithCreate withExistingResourceGroup(String resourceGroupName);
         }
+
         /**
          * The stage of the PartnerDestination definition which contains all the minimum required properties for the
          * resource to be created, but also allows for any other optional properties to be specified.
@@ -203,6 +207,7 @@ public interface PartnerDestination {
              */
             PartnerDestination create(Context context);
         }
+
         /** The stage of the PartnerDestination definition allowing to specify tags. */
         interface WithTags {
             /**
@@ -213,6 +218,7 @@ public interface PartnerDestination {
              */
             WithCreate withTags(Map<String, String> tags);
         }
+
         /** The stage of the PartnerDestination definition allowing to specify partnerRegistrationImmutableId. */
         interface WithPartnerRegistrationImmutableId {
             /**
@@ -224,6 +230,7 @@ public interface PartnerDestination {
              */
             WithCreate withPartnerRegistrationImmutableId(UUID partnerRegistrationImmutableId);
         }
+
         /** The stage of the PartnerDestination definition allowing to specify endpointServiceContext. */
         interface WithEndpointServiceContext {
             /**
@@ -235,6 +242,7 @@ public interface PartnerDestination {
              */
             WithCreate withEndpointServiceContext(String endpointServiceContext);
         }
+
         /** The stage of the PartnerDestination definition allowing to specify expirationTimeIfNotActivatedUtc. */
         interface WithExpirationTimeIfNotActivatedUtc {
             /**
@@ -249,6 +257,7 @@ public interface PartnerDestination {
              */
             WithCreate withExpirationTimeIfNotActivatedUtc(OffsetDateTime expirationTimeIfNotActivatedUtc);
         }
+
         /** The stage of the PartnerDestination definition allowing to specify activationState. */
         interface WithActivationState {
             /**
@@ -259,6 +268,7 @@ public interface PartnerDestination {
              */
             WithCreate withActivationState(PartnerDestinationActivationState activationState);
         }
+
         /** The stage of the PartnerDestination definition allowing to specify endpointBaseUrl. */
         interface WithEndpointBaseUrl {
             /**
@@ -269,6 +279,7 @@ public interface PartnerDestination {
              */
             WithCreate withEndpointBaseUrl(String endpointBaseUrl);
         }
+
         /** The stage of the PartnerDestination definition allowing to specify messageForActivation. */
         interface WithMessageForActivation {
             /**
@@ -281,6 +292,7 @@ public interface PartnerDestination {
             WithCreate withMessageForActivation(String messageForActivation);
         }
     }
+
     /**
      * Begins update for the PartnerDestination resource.
      *
@@ -305,6 +317,7 @@ public interface PartnerDestination {
          */
         PartnerDestination apply(Context context);
     }
+
     /** The PartnerDestination update stages. */
     interface UpdateStages {
         /** The stage of the PartnerDestination update allowing to specify tags. */
@@ -318,6 +331,7 @@ public interface PartnerDestination {
             Update withTags(Map<String, String> tags);
         }
     }
+
     /**
      * Refreshes the resource to sync with Azure.
      *
