+++ sdk/eventgrid/azure-resourcemanager-eventgrid/src/main/java/com/azure/resourcemanager/eventgrid/models/Topic.java
@@ -213,11 +213,13 @@ public interface Topic {
             DefinitionStages.WithResourceGroup,
             DefinitionStages.WithCreate {
     }
+
     /** The Topic definition stages. */
     interface DefinitionStages {
         /** The first stage of the Topic definition. */
         interface Blank extends WithLocation {
         }
+
         /** The stage of the Topic definition allowing to specify location. */
         interface WithLocation {
             /**
@@ -236,6 +238,7 @@ public interface Topic {
              */
             WithResourceGroup withRegion(String location);
         }
+
         /** The stage of the Topic definition allowing to specify parent resource. */
         interface WithResourceGroup {
             /**
@@ -246,6 +249,7 @@ public interface Topic {
              */
             WithCreate withExistingResourceGroup(String resourceGroupName);
         }
+
         /**
          * The stage of the Topic definition which contains all the minimum required properties for the resource to be
          * created, but also allows for any other optional properties to be specified.
@@ -279,6 +283,7 @@ public interface Topic {
              */
             Topic create(Context context);
         }
+
         /** The stage of the Topic definition allowing to specify tags. */
         interface WithTags {
             /**
@@ -289,6 +294,7 @@ public interface Topic {
              */
             WithCreate withTags(Map<String, String> tags);
         }
+
         /** The stage of the Topic definition allowing to specify sku. */
         interface WithSku {
             /**
@@ -299,6 +305,7 @@ public interface Topic {
              */
             WithCreate withSku(ResourceSku sku);
         }
+
         /** The stage of the Topic definition allowing to specify identity. */
         interface WithIdentity {
             /**
@@ -309,6 +316,7 @@ public interface Topic {
              */
             WithCreate withIdentity(IdentityInfo identity);
         }
+
         /** The stage of the Topic definition allowing to specify kind. */
         interface WithKind {
             /**
@@ -319,6 +327,7 @@ public interface Topic {
              */
             WithCreate withKind(ResourceKind kind);
         }
+
         /** The stage of the Topic definition allowing to specify extendedLocation. */
         interface WithExtendedLocation {
             /**
@@ -329,6 +338,7 @@ public interface Topic {
              */
             WithCreate withExtendedLocation(ExtendedLocation extendedLocation);
         }
+
         /** The stage of the Topic definition allowing to specify eventTypeInfo. */
         interface WithEventTypeInfo {
             /**
@@ -342,6 +352,7 @@ public interface Topic {
              */
             WithCreate withEventTypeInfo(EventTypeInfo eventTypeInfo);
         }
+
         /** The stage of the Topic definition allowing to specify minimumTlsVersionAllowed. */
         interface WithMinimumTlsVersionAllowed {
             /**
@@ -353,6 +364,7 @@ public interface Topic {
              */
             WithCreate withMinimumTlsVersionAllowed(TlsVersion minimumTlsVersionAllowed);
         }
+
         /** The stage of the Topic definition allowing to specify inputSchema. */
         interface WithInputSchema {
             /**
@@ -365,6 +377,7 @@ public interface Topic {
              */
             WithCreate withInputSchema(InputSchema inputSchema);
         }
+
         /** The stage of the Topic definition allowing to specify inputSchemaMapping. */
         interface WithInputSchemaMapping {
             /**
@@ -379,6 +392,7 @@ public interface Topic {
              */
             WithCreate withInputSchemaMapping(InputSchemaMapping inputSchemaMapping);
         }
+
         /** The stage of the Topic definition allowing to specify publicNetworkAccess. */
         interface WithPublicNetworkAccess {
             /**
@@ -394,6 +408,7 @@ public interface Topic {
              */
             WithCreate withPublicNetworkAccess(PublicNetworkAccess publicNetworkAccess);
         }
+
         /** The stage of the Topic definition allowing to specify inboundIpRules. */
         interface WithInboundIpRules {
             /**
@@ -406,6 +421,7 @@ public interface Topic {
              */
             WithCreate withInboundIpRules(List<InboundIpRule> inboundIpRules);
         }
+
         /** The stage of the Topic definition allowing to specify disableLocalAuth. */
         interface WithDisableLocalAuth {
             /**
@@ -420,6 +436,7 @@ public interface Topic {
              */
             WithCreate withDisableLocalAuth(Boolean disableLocalAuth);
         }
+
         /** The stage of the Topic definition allowing to specify dataResidencyBoundary. */
         interface WithDataResidencyBoundary {
             /**
@@ -431,6 +448,7 @@ public interface Topic {
             WithCreate withDataResidencyBoundary(DataResidencyBoundary dataResidencyBoundary);
         }
     }
+
     /**
      * Begins update for the Topic resource.
      *
@@ -464,6 +482,7 @@ public interface Topic {
          */
         Topic apply(Context context);
     }
+
     /** The Topic update stages. */
     interface UpdateStages {
         /** The stage of the Topic update allowing to specify tags. */
@@ -476,6 +495,7 @@ public interface Topic {
              */
             Update withTags(Map<String, String> tags);
         }
+
         /** The stage of the Topic update allowing to specify identity. */
         interface WithIdentity {
             /**
@@ -486,6 +506,7 @@ public interface Topic {
              */
             Update withIdentity(IdentityInfo identity);
         }
+
         /** The stage of the Topic update allowing to specify sku. */
         interface WithSku {
             /**
@@ -496,6 +517,7 @@ public interface Topic {
              */
             Update withSku(ResourceSku sku);
         }
+
         /** The stage of the Topic update allowing to specify publicNetworkAccess. */
         interface WithPublicNetworkAccess {
             /**
@@ -512,6 +534,7 @@ public interface Topic {
              */
             Update withPublicNetworkAccess(PublicNetworkAccess publicNetworkAccess);
         }
+
         /** The stage of the Topic update allowing to specify inboundIpRules. */
         interface WithInboundIpRules {
             /**
@@ -524,6 +547,7 @@ public interface Topic {
              */
             Update withInboundIpRules(List<InboundIpRule> inboundIpRules);
         }
+
         /** The stage of the Topic update allowing to specify minimumTlsVersionAllowed. */
         interface WithMinimumTlsVersionAllowed {
             /**
@@ -535,6 +559,7 @@ public interface Topic {
              */
             Update withMinimumTlsVersionAllowed(TlsVersion minimumTlsVersionAllowed);
         }
+
         /** The stage of the Topic update allowing to specify disableLocalAuth. */
         interface WithDisableLocalAuth {
             /**
@@ -549,6 +574,7 @@ public interface Topic {
              */
             Update withDisableLocalAuth(Boolean disableLocalAuth);
         }
+
         /** The stage of the Topic update allowing to specify dataResidencyBoundary. */
         interface WithDataResidencyBoundary {
             /**
@@ -559,6 +585,7 @@ public interface Topic {
              */
             Update withDataResidencyBoundary(DataResidencyBoundary dataResidencyBoundary);
         }
+
         /** The stage of the Topic update allowing to specify eventTypeInfo. */
         interface WithEventTypeInfo {
             /**
@@ -570,6 +597,7 @@ public interface Topic {
             Update withEventTypeInfo(EventTypeInfo eventTypeInfo);
         }
     }
+
     /**
      * Refreshes the resource to sync with Azure.
      *
