+++ sdk/eventgrid/azure-resourcemanager-eventgrid/src/main/java/com/azure/resourcemanager/eventgrid/models/TopicEventSubscriptions.java
@@ -7,6 +7,7 @@ package com.azure.resourcemanager.eventgrid.models;
 import com.azure.core.http.rest.PagedIterable;
 import com.azure.core.http.rest.Response;
 import com.azure.core.util.Context;
+import com.azure.resourcemanager.eventgrid.fluent.models.EventSubscriptionInner;
 
 /** Resource collection API of TopicEventSubscriptions. */
 public interface TopicEventSubscriptions {
@@ -78,6 +79,50 @@ public interface TopicEventSubscriptions {
     EventSubscription get(String resourceGroupName, String topicName, String eventSubscriptionName);
 
     /**
+     * Create or update an event subscription to a topic.
+     *
+     * <p>Asynchronously creates a new event subscription or updates an existing event subscription.
+     *
+     * @param resourceGroupName The name of the resource group within the user's subscription.
+     * @param topicName Name of the domain topic.
+     * @param eventSubscriptionName Name of the event subscription to be created. Event subscription names must be
+     *     between 3 and 100 characters in length and use alphanumeric letters only.
+     * @param eventSubscriptionInfo Event subscription properties containing the destination and filter information.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return event Subscription.
+     */
+    EventSubscription createOrUpdate(
+        String resourceGroupName,
+        String topicName,
+        String eventSubscriptionName,
+        EventSubscriptionInner eventSubscriptionInfo);
+
+    /**
+     * Create or update an event subscription to a topic.
+     *
+     * <p>Asynchronously creates a new event subscription or updates an existing event subscription.
+     *
+     * @param resourceGroupName The name of the resource group within the user's subscription.
+     * @param topicName Name of the domain topic.
+     * @param eventSubscriptionName Name of the event subscription to be created. Event subscription names must be
+     *     between 3 and 100 characters in length and use alphanumeric letters only.
+     * @param eventSubscriptionInfo Event subscription properties containing the destination and filter information.
+     * @param context The context to associate with this operation.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return event Subscription.
+     */
+    EventSubscription createOrUpdate(
+        String resourceGroupName,
+        String topicName,
+        String eventSubscriptionName,
+        EventSubscriptionInner eventSubscriptionInfo,
+        Context context);
+
+    /**
      * Delete an event subscription for a topic.
      *
      * <p>Delete an existing event subscription for a topic.
@@ -107,6 +152,48 @@ public interface TopicEventSubscriptions {
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      */
     void delete(String resourceGroupName, String topicName, String eventSubscriptionName, Context context);
+
+    /**
+     * Update an event subscription for a topic.
+     *
+     * <p>Update an existing event subscription for a topic.
+     *
+     * @param resourceGroupName The name of the resource group within the user's subscription.
+     * @param topicName Name of the domain.
+     * @param eventSubscriptionName Name of the event subscription to be updated.
+     * @param eventSubscriptionUpdateParameters Updated event subscription information.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return event Subscription.
+     */
+    EventSubscription update(
+        String resourceGroupName,
+        String topicName,
+        String eventSubscriptionName,
+        EventSubscriptionUpdateParameters eventSubscriptionUpdateParameters);
+
+    /**
+     * Update an event subscription for a topic.
+     *
+     * <p>Update an existing event subscription for a topic.
+     *
+     * @param resourceGroupName The name of the resource group within the user's subscription.
+     * @param topicName Name of the domain.
+     * @param eventSubscriptionName Name of the event subscription to be updated.
+     * @param eventSubscriptionUpdateParameters Updated event subscription information.
+     * @param context The context to associate with this operation.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return event Subscription.
+     */
+    EventSubscription update(
+        String resourceGroupName,
+        String topicName,
+        String eventSubscriptionName,
+        EventSubscriptionUpdateParameters eventSubscriptionUpdateParameters,
+        Context context);
 
     /**
      * Get full URL of an event subscription for topic.
@@ -177,64 +264,4 @@ public interface TopicEventSubscriptions {
      */
     PagedIterable<EventSubscription> list(
         String resourceGroupName, String topicName, String filter, Integer top, Context context);
-
-    /**
-     * Get an event subscription of a topic.
-     *
-     * <p>Get properties of an event subscription of a topic.
-     *
-     * @param id the resource ID.
-     * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
-     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return properties of an event subscription of a topic along with {@link Response}.
-     */
-    EventSubscription getById(String id);
-
-    /**
-     * Get an event subscription of a topic.
-     *
-     * <p>Get properties of an event subscription of a topic.
-     *
-     * @param id the resource ID.
-     * @param context The context to associate with this operation.
-     * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
-     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return properties of an event subscription of a topic along with {@link Response}.
-     */
-    Response<EventSubscription> getByIdWithResponse(String id, Context context);
-
-    /**
-     * Delete an event subscription for a topic.
-     *
-     * <p>Delete an existing event subscription for a topic.
-     *
-     * @param id the resource ID.
-     * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
-     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     */
-    void deleteById(String id);
-
-    /**
-     * Delete an event subscription for a topic.
-     *
-     * <p>Delete an existing event subscription for a topic.
-     *
-     * @param id the resource ID.
-     * @param context The context to associate with this operation.
-     * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
-     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     */
-    void deleteByIdWithResponse(String id, Context context);
-
-    /**
-     * Begins definition for a new EventSubscription resource.
-     *
-     * @param name resource name.
-     * @return the first stage of the new EventSubscription definition.
-     */
-    EventSubscription.DefinitionStages.Blank define(String name);
 }
